Remove Analytics-specific quirks out of the core module setup component.
