+++ assets/js/components/setup/DefaultModuleSetup.tsx
@@ -17,12 +17,6 @@
  */
 
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-import { useMount } from 'react-use';
-
-/**
  * WordPress dependencies
  */
 import { Fragment } from '@wordpress/element';
@@ -32,14 +26,8 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import { type Select, useSelect } from 'googlesitekit-data';
-import ProgressIndicator from '@/js/components/ProgressIndicator';
-import ExitSetup from '@/js/components/setup/ExitSetup';
 import { CORE_MODULES } from '@/js/googlesitekit/modules/datastore/constants';
-import { useFeature } from '@/js/hooks/useFeature';
-import useQueryArg from '@/js/hooks/useQueryArg';
-import useViewContext from '@/js/hooks/useViewContext';
 import { Cell, Grid, Row } from '@/js/material-components';
-import { MODULE_SLUG_ANALYTICS_4 } from '@/js/modules/analytics-4/constants';
 import { useFinishSetup, useModuleSetupTracking } from './hooks';
 import SetupFooter from './SetupFooter';
 import SetupHeader from './SetupHeader';
@@ -48,126 +36,55 @@ import type { ModuleSetupLayoutProps, ModuleWithSetupComponent } from './types';
 export default function DefaultModuleSetup( {
 	moduleSlug,
 }: ModuleSetupLayoutProps ) {
-	const viewContext = useViewContext();
-	const setupFlowRefreshEnabled = useFeature( 'setupFlowRefresh' );
-	const [ showProgress ] = useQueryArg( 'showProgress' );
-
-	const isInitialSetupFlow =
-		setupFlowRefreshEnabled &&
-		moduleSlug === MODULE_SLUG_ANALYTICS_4 &&
-		showProgress === 'true';
-
 	const module = useSelect(
 		( select: Select ): ModuleWithSetupComponent | undefined =>
 			select( CORE_MODULES ).getModule( moduleSlug ),
 		[ moduleSlug ]
 	);
 
-	const finishSetup = useFinishSetup(
-		moduleSlug,
-		isInitialSetupFlow
-			? {
-					gaTrackingEventArgs: {
-						category: `${ viewContext }_setup`,
-						action: 'setup_flow_v3_complete_analytics_step',
-					},
-			  }
-			: {}
-	);
+	const finishSetup = useFinishSetup( moduleSlug );
 
-	const { trackCancel: onCancelButtonClick } = useModuleSetupTracking(
-		moduleSlug,
-		isInitialSetupFlow
-			? {
-					viewGATrackingEventArgs: {
-						category: `${ viewContext }_setup`,
-						action: 'setup_flow_v3_view_analytics_step',
-					},
-			  }
-			: {}
-	);
-
-	// Add the initial setup class to the body when the component mounts.
-	useMount( () => {
-		if ( isInitialSetupFlow ) {
-			global.document.body.classList.add( 'googlesitekit-setup-flow' );
-		}
-	} );
-
+	const { trackCancel: onCancelButtonClick } =
+		useModuleSetupTracking( moduleSlug );
 	if ( ! module?.SetupComponent ) {
 		return null;
 	}
 
 	const { SetupComponent } = module;
 
-	const setupMainContent = (
+	return (
 		<Fragment>
-			{ isInitialSetupFlow && (
-				<ProgressIndicator currentSegment={ 3 } totalSegments={ 6 } />
-			) }
-			<section className="googlesitekit-setup__wrapper">
+			<SetupHeader />
+			<div className="googlesitekit-setup">
 				<Grid>
 					<Row>
 						<Cell size={ 12 }>
-							{ ! isInitialSetupFlow && (
-								<p className="googlesitekit-setup__intro-title">
-									{ __(
-										'Connect Service',
-										'google-site-kit'
-									) }
-								</p>
-							) }
-							<SetupComponent
-								module={ module }
-								finishSetup={ finishSetup }
-							/>
+							<section className="googlesitekit-setup__wrapper">
+								<Grid>
+									<Row>
+										<Cell size={ 12 }>
+											<p className="googlesitekit-setup__intro-title">
+												{ __(
+													'Connect Service',
+													'google-site-kit'
+												) }
+											</p>
+											<SetupComponent
+												module={ module }
+												finishSetup={ finishSetup }
+											/>
+										</Cell>
+									</Row>
+								</Grid>
+								<SetupFooter
+									moduleSlug={ moduleSlug }
+									finishSetup={ finishSetup }
+									onCancel={ onCancelButtonClick }
+								/>
+							</section>
 						</Cell>
 					</Row>
 				</Grid>
-
-				{ ! isInitialSetupFlow && (
-					<SetupFooter
-						moduleSlug={ moduleSlug }
-						finishSetup={ finishSetup }
-						onCancel={ onCancelButtonClick }
-					/>
-				) }
-			</section>
-		</Fragment>
-	);
-
-	return (
-		<Fragment>
-			<SetupHeader>
-				{ isInitialSetupFlow && (
-					<ExitSetup
-						gaTrackingEventArgs={ {
-							category: `${ viewContext }_setup`,
-							label: moduleSlug,
-						} }
-					/>
-				) }
-			</SetupHeader>
-			<div
-				className={ classnames(
-					'googlesitekit-setup',
-					...( isInitialSetupFlow
-						? [
-								'googlesitekit-initial-setup',
-								`googlesitekit-initial-setup--${ moduleSlug }`,
-						  ]
-						: [] )
-				) }
-			>
-				{ isInitialSetupFlow ? (
-					setupMainContent
-				) : (
-					<Grid>
-						<Row>
-							<Cell size={ 12 }>{ setupMainContent }</Cell>
-						</Row>
-					</Grid>
-				) }
 			</div>
 		</Fragment>
 	);
