+++ assets/js/modules/analytics-4/index.js
@@ -27,7 +27,7 @@ import { __ } from '@wordpress/i18n';
 import AnalyticsIcon from '@/svg/graphics/analytics.svg';
 import DashboardMainEffectComponent from './components/DashboardMainEffectComponent';
 import { SettingsEdit, SettingsView } from './components/settings';
-import { SetupMain } from './components/setup';
+import { SetupLayout, SetupMain } from './components/setup';
 import { MODULE_SLUG_ANALYTICS_4 } from './constants';
 import { MODULES_ANALYTICS_4 } from './datastore/constants';
 
@@ -40,6 +40,7 @@ export function registerModule( modules ) {
 		storeName: MODULES_ANALYTICS_4,
 		SettingsEditComponent: SettingsEdit,
 		SettingsViewComponent: SettingsView,
+		SetupLayout,
 		SetupComponent: SetupMain,
 		DashboardMainEffectComponent,
 		Icon: AnalyticsIcon,
